requirements.txt
pycryptodome>=3.19.0
pytest>=7.0.0
jupyter>=1.0.0
ipython>=8.0.0
black>=23.0.0
flake8>=6.0.0
matplotlib>=3.7.0
numpy>=1.24.0